backend/requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==21.2.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.12.0
sqlalchemy-utils==0.41.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
google-auth==2.29.0
google-auth-oauthlib==1.1.0
pydantic==2.4.2
google-auth-httplib2==0.1.1
requests==2.31.0
pyJWT==2.8.0

# Data Processing
pandas==2.2.2
pydantic-settings==2.0.3
python-multipart==0.0.7
openpyxl==3.1.2
numpy==1.26.4


# HTTP & Networking
httpx==0.25.0
slowapi==0.1.8

# Monitoring & Logging
structlog==23.2.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-env==1.1.1
pytest-sugar==0.9.7
coverage==7.3.2
pytest-benchmark==4.0.0
faker==19.3.1
freezegun==1.2.2

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1